perf(pages/progress-connect): remove unused telemetry
