+++ src/pages/popup/lib/context.tsx
@@ -5,19 +5,16 @@ import {
   type PopupToBackgroundMessage,
   type BackgroundToPopupMessage,
 } from '@/shared/messages';
-import {
-  TelemetryContextProvider,
-  useBrowser,
-  useTranslation,
-} from '@/pages/shared/lib/context';
+import { useBrowser, useTranslation } from '@/pages/shared/lib/context';
+import { TelemetryContextProvider } from '@/pages/shared/lib/telemetry';
 import { dispatch } from './store';
 
 export {
   useBrowser,
   useBrowserInfo,
   useTranslation,
-  useTelemetry,
 } from '@/pages/shared/lib/context';
+export { useTelemetry } from '@/pages/shared/lib/telemetry';
 
 export function WaitForStateLoad({ children }: React.PropsWithChildren) {
   const t = useTranslation();
